--- a/lib/matplotlib/cm.py
+++ b/lib/matplotlib/cm.py
@@ -735,5 +735,6 @@ def _ensure_cmap(cmap):
     cmap_name = cmap if cmap is not None else mpl.rcParams["image.cmap"]
     # use check_in_list to ensure type stability of the exception raised by
     # the internal usage of this (ValueError vs KeyError)
-    _api.check_in_list(list(_colormaps), cmap=cmap_name)
+    if cmap_name not in _colormaps:
+        _api.check_in_list(sorted(_colormaps), cmap=cmap_name)
     return mpl.colormaps[cmap_name]