--- a/lib/matplotlib/colors.py
+++ b/lib/matplotlib/colors.py
@@ -1390,6 +1390,10 @@ class Normalize:
     def autoscale_None(self, A):
         """If vmin or vmax are not set, use the min/max of *A* to set them."""
         A = np.asanyarray(A)
+
+        if isinstance(A, np.ma.MaskedArray) and A.mask is False:
+            A = A.data
+
         if self.vmin is None and A.size:
             self.vmin = A.min()
         if self.vmax is None and A.size: