--- a/lib/matplotlib/colors.py
+++ b/lib/matplotlib/colors.py
@@ -1391,8 +1391,10 @@ class Normalize:
         """If vmin or vmax are not set, use the min/max of *A* to set them."""
         A = np.asanyarray(A)
 
-        if isinstance(A, np.ma.MaskedArray) and not A.mask:
-            A = A.data
+        if isinstance(A, np.ma.MaskedArray):
+            # we need to make the distinction between an array, False, np.bool_(False)
+            if A.mask is False or not A.mask.shape:
+                A = A.data
 
         if self.vmin is None and A.size:
             self.vmin = A.min()