--- a/lib/matplotlib/tests/test_cbook.py
+++ b/lib/matplotlib/tests/test_cbook.py
@@ -15,7 +15,7 @@ import pytest
 
 from matplotlib import _api, cbook
 import matplotlib.colors as mcolors
-from matplotlib.cbook import delete_masked_points, strip_math
+from matplotlib.cbook import delete_masked_points
 
 
 class Test_delete_masked_points: