## PR summary

In this PR we apply some micro optimizations to improve the imshow performance. Benchmark:
```
Mean +- std dev: [main] 7.34 ms +- 0.07 ms -> [pr] 6.67 ms +- 0.11 ms: 1.10x faster
```

Notes:

- There is a fast path for the `MaskedArray` in case the mask is `False` (e.g. images without a NaN or Inf value)
- The `np.ma.masked_invalid` undoes the `shrink_mask` from `np.ma.masked_where`, but matplotlib applies it again. So it is faster to call `np.ma.masked_where` directly
- Another possible optimization is in `Normalize.autoscale_None`. There both the setting of `vmin` and `vmax` triggers `_changed`. It is faster to first set vmin and vmax (without triggering _changed) and then calling _changed. This would require a new method `set_vmin_vmax` or some other mechanism to avoid triggering `_changed` twice.

Benchmark script:
```
import matplotlib
# print(matplotlib)
import pyperf

setup = """
import matplotlib.pyplot as plt
import numpy as np

n=10
im=np.random.rand(100,100)

def go():
    for ii in range(n):
        plt.figure(1)
        plt.imshow(im)
"""

runner = pyperf.Runner()
runner.timeit(name="mpl", stmt="go()", setup=setup)
```

## PR checklist
<!-- Please mark any checkboxes that do not apply to this PR as [N/A].-->

- [N/A] "closes #0000" is in the body of the PR description to [link the related issue](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue)
- [ ] new and changed code is [tested](https://matplotlib.org/devdocs/devel/testing.html)
- [N/A] *Plotting related* features are demonstrated in an [example](https://matplotlib.org/devdocs/devel/documenting_mpl.html#writing-examples-and-tutorials)
- [N/A] *New Features* and *API Changes* are noted with a [directive and release note](https://matplotlib.org/devdocs/devel/coding_guide.html#new-features-and-api-changes)
- [N/A] Documentation complies with [general](https://matplotlib.org/devdocs/devel/documenting_mpl.html#writing-rest-pages) and [docstring](https://matplotlib.org/devdocs/devel/documenting_mpl.html#writing-docstrings) guidelines

<!--
Thank you so much for your PR!  To help us review your contribution, please
consider the following points:

- A development guide is available at https://matplotlib.org/devdocs/devel/index.html.

- Help with git and github is available at https://matplotlib.org/devdocs/devel/development_workflow.html

- Create a separate branch for your changes and open the PR from this branch. Please avoid working on `main`.

- The PR title should summarize the changes, for example "Raise ValueError on
  non-numeric input to set_xlim".  Avoid non-descriptive titles such as
  "Addresses issue #8576".

- The summary should provide at least 1-2 sentences describing the pull request
  in detail (Why is this change required?  What problem does it solve?) and
  link to any relevant issues.

- If you are contributing fixes to docstrings, please pay attention to
  https://matplotlib.org/stable/devel/documenting_mpl.html#formatting-conventions.  In particular,
  note the difference between using single backquotes, double backquotes, and
  asterisks in the markup.

We understand that PRs can sometimes be overwhelming, especially as the
reviews start coming in.  Please let us know if the reviews are unclear or
the recommended next step seems overly demanding, if you would like help in
addressing a reviewer's comments, or if you have been waiting too long to hear
back on your PR.
-->