@eendebakpt 2023-07-17
Coverage was complaining because the number of lines of code was reduced. I added a few simple tests for `strip_math` to make ci happy.