@tacaswell 2023-07-17
Is `stip_math` otherwise implicated is this change and does the new test test something that we do not already test?

Some of the codecov failures sort them selves out when all of the CI jobs upload their results and are processed (as some of the tests are platform specific)