@eendebakpt 2023-07-18
> Is `stip_math` otherwise implicated is this change and does the new test test something that we do not already test?
> 
> Some of the codecov failures sort them selves out when all of the CI jobs upload their results and are processed (as some of the tests are platform specific)

The `strip_math` is completely unrelated to this PR. Coverage showed the method was not in the unit testing (at least not explicit), so I picked it as an easy way to increase coverage. If you want I can remove the tests again and see what happens to the CI.