@jklymak 2023-07-18
Maybe directly testing `strip_math` is OK, but it points to whatever code calls `strip_math` as not being tested either.  It gets used in `LogFormatter`, which I'd have thought was thoroughly tested, but maybe not?